fix: cleanup after rebase
