--- AMW_business/src/main/java/ch/puzzle/itc/mobiliar/business/resourcegroup/boundary/GetResourceService.java
@@ -1,43 +1,0 @@
-/*
- * AMW - Automated Middleware allows you to manage the configurations of
- * your Java EE applications on an unlimited number of different environments
- * with various versions, including the automated deployment of those apps.
- * Copyright (C) 2013-2026 by Puzzle ITC
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package ch.puzzle.itc.mobiliar.business.resourcegroup.boundary;
-
-import ch.puzzle.itc.mobiliar.business.resourcegroup.entity.ResourceEntity;
-import ch.puzzle.itc.mobiliar.common.exception.ResourceNotFoundException;
-
-import javax.ejb.Stateless;
-import javax.inject.Inject;
-
-@Stateless
-public class GetResourceService implements GetResourceUseCase {
-
-    @Inject
-    private ResourceLocator resourceLocator;
-
-    @Override
-    public ResourceEntity getWithGroupAndRelatedResources(Integer resourceId) throws ResourceNotFoundException {
-        ResourceEntity resource = resourceLocator.getResourceWithGroupAndRelatedResources(resourceId);
-        if (resource == null) {
-            throw new ResourceNotFoundException("Resource with id " + resourceId + " not found");
-        }
-        return resource;
-    }
-}
