+++ AMW_business/src/main/java/ch/puzzle/itc/mobiliar/business/resourcegroup/boundary/GetResourceUseCase.java
@@ -1,6 +1,7 @@
 package ch.puzzle.itc.mobiliar.business.resourcegroup.boundary;
 
 import ch.puzzle.itc.mobiliar.business.resourcegroup.entity.ResourceEntity;
+import ch.puzzle.itc.mobiliar.common.exception.ResourceNotFoundException;
 
 public interface GetResourceUseCase {
